refactor(builder): extract build_agent_inner to agent_inner.rs (dirge-4y4l stage 11c)
Final structural stage for the builder god-file. Moves the core agent
constructor out of builder/mod.rs, leaving mod.rs as pure wiring.

- agent_inner.rs (570 LOC) — build_agent_inner (~516-LOC constructor that
  wires the rig Agent: preamble + tool registry + provider model under the
  wide DI signature) + the plugin `hookify` wrapper it applies to every
  tool. Composes the preamble helpers + build_session_search_tool via the
  parent's re-exports.
- mod.rs (33 LOC, was 586) — module wiring + build_session_search_tool
  (the one helper shared by agent_inner and loop_tools) only. Imports
  pruned to that helper's needs.
- reminder_tests.rs — gains explicit imports for the crate types it
  constructs (Cli/Config/ToolCache/Sandbox/resolve_family), previously
  reached via the parent module's now-relocated `use` statements.

Also fixes a stray from stage 11b (#306): wrap_mcp_tools lost its
`#[cfg(feature = "mcp")]` gate and its doc comment (extraction started at
the fn line, not the attr/doc), and the orphaned attr wrongly gated
`mod reminder_tests` on mcp. Restored the gate + doc on wrap_mcp_tools in
loop_tools.rs and removed the orphan. (No CI impact: default and
windows-default both include mcp.)

End state: builder.rs 1835 -> builder/{mod 33, agent_inner 570,
loop_tools 519, preamble 89, reminder_tests 697}.
-D warnings clean default + windows-default; 2374 tests green.

Refs dirge-1g9q
